useDebounce moved into 'hook' folder. Removed from movieList and peopleList
