+++ src/hooks/useDebounce.js
@@ -1,0 +1,22 @@
+import { useDispatch } from "react-redux";
+import { setLoadingState } from "../Redux_store/settingSlice";
+import { useEffect, useState } from "react";
+
+const useDebounce = (value, delay) => {
+  const [debouncedValue, setDebouncedValue] = useState(value);
+  const dispatch = useDispatch();
+
+  useEffect(() => {
+    dispatch(setLoadingState("loading"));
+    const handler = setTimeout(() => {
+      setDebouncedValue(value);
+    }, delay);
+    return () => {
+      clearTimeout(handler);
+    };
+  }, [value, delay]);
+
+  return debouncedValue;
+};
+
+export default useDebounce;
